test: merge default-session-store.test.ts into session-store.test.ts
One source file → one test file.
